requirements.txt
Django==1.9.2
argparse==1.2.1
django-registration-redux==1.3
wsgiref==0.1.2